requirements.txt
streamlit
requests
langchain
groq
python-dotenv